+++ bootique-cayenne42/src/main/java/io/bootique/cayenne/v42/commitlog/MappedCommitLogListener.java
@@ -1,0 +1,43 @@
+/*
+ * Licensed to ObjectStyle LLC under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ObjectStyle LLC licenses
+ * this file to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package io.bootique.cayenne.v42.commitlog;
+
+import org.apache.cayenne.commitlog.CommitLogListener;
+
+/**
+ * @since 3.0.M1
+ */
+public class MappedCommitLogListener {
+
+    private final CommitLogListener listener;
+    private final boolean includeInTransaction;
+
+    public MappedCommitLogListener(CommitLogListener listener, boolean includeInTransaction) {
+        this.listener = listener;
+        this.includeInTransaction = includeInTransaction;
+    }
+
+    public CommitLogListener getListener() {
+        return listener;
+    }
+
+    public boolean isIncludeInTransaction() {
+        return includeInTransaction;
+    }
+}